Better reset level overlay location
